Update README.mdHtml
HTML, or HyperText Markup Language, is the standard markup language used to create web pages. It provides the structure and content of a webpage, defining elements like headings, paragraphs, and links. HTML uses a system of tags (e.g., <p>, <h1>, <a>) to mark up text and other content, telling browsers how to display the information. 
Here's a more detailed breakdown:
Key Concepts:
Markup Language:
HTML is a markup language, meaning it uses tags to annotate text and define how it should be presented. 
Structure:
HTML defines the structure of a webpage, outlining the different sections and elements that make up the page's content. 
Elements:
HTML uses elements (defined by tags) to represent different parts of the page, such as headings, paragraphs, images, and links. 
Tags:
Tags are keywords enclosed in angle brackets (< and >). They indicate the beginning and end of an element. 
Attributes:
Tags can have attributes that provide additional information about the element, like the source of an image or the link destination.